move config file to app dir
